BUG-016 Fix: Sync Standard/Wide Layout Toggles
The standard/wide mode toggle under the account settings and the toggle on the left hand ribbon were out of sync with one another. Implemented fix makes it so when there is an update to either one of the toggles, the other will follow suit. Also fixed an error with passing the commit tests for the previous fix which resolved the light/dark mode toggles being out of sync