Traceability View Framework First Iteration
This is the first iteration of the traceability view. These changes contain updates to src\components\base\AppSidebar.tsx which implement a dropdown menu to navigate to the different view included in the new traceability page. The src\app\(protected)\traceability\page.tsx file contains the new page for traceability, which at this point is hooked up with sample data. This page will have 3 views, Matrix View which will contain all the test requirements contained in a project, Hierarchy View which will show the hierarchy structure of all the requirements within a project, and Test Requirements, which will be an analytical break down of the test results.